Move vast files to new directory
